simple-python-service/requirements.txt

sqs_py
pySqsListener
boto3
